feat(FamilyCard): adjust FamilyCard to accept PersonCard(s) as children
